biiig update added debug argument (-d) audio is now a feature (included by default) removed winres and put embed_resource in its place ui now has its own function called after render moved npc folders and tile textures removed extra traits in util redid map, warp, npc, and script files added some new maps and removed unused ones

+++ src/gui/battle/pokemon_gui.rs
@@ -99,7 +99,7 @@ impl PokemonGui for PlayerPokemonGui {
 			self.level.render();
 			self.health_text.render();
 			self.health_bar.render();
-			draw_rect(Self::EXP_COLOR, self.panel.x + 32.0, self.panel.y + 33.0, (self.exp_width * 64.0) as u32, 2);
+			draw_rect(Self::EXP_COLOR, self.panel.x + 32.0, self.panel.y + 33.0, self.exp_width * 64.0, 2.0);
 		}		
 	}
